# event-reservation-api-swagger-v2.json
{
  "swagger":"2.0",
  "info":{
    "description":"\"Allows you to View Availability, Book, Change & Cancel Reservations.\"",
    "version":"1.0.0",
    "title":"Event Reservation REST API",
    "contact":{
      "name":"Kevin Gomes",
      "url":"http://kgomes.info/",
      "email":"kevingomes17@gmail.com"
    },
    "license":{
      "name":"Apache License Version 2.0",
      "url":"https://www.apache.org/licenses/LICENSE-2.0\""
    }
  },
  "host":"localhost:8080",
  "schemes": ["http"],
  "basePath":"/",
  "tags":[
    {
      "name":"reservation-controller",
      "description":"Reservation Controller"
    }
  ],
  "paths":{
    "/reservation/available-dates":{
      "get":{
        "tags":[
          "reservation-controller"
        ],
        "summary":"Check availability",
        "description":"Check availability either by specifying a date range. Or omitting the Date Range will fetch availability for the next 30 days.",
        "operationId":"getAvailableDatesForUsingGET",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "responses":{
          "200":{
            "description":"Success",
            "schema":{
              "$ref":"#/definitions/AvailabilitySearchRes"
            }
          }
        }
      }
    },
    "/reservation/book":{
      "post":{
        "tags":[
          "reservation-controller"
        ],
        "summary":"Book Reservation",
        "description":"Books a reservation after all the validation criteria are met. Handles Race Condition.",
        "operationId":"bookUsingPOST",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"body",
            "name":"req",
            "description":"req",
            "required":true,
            "schema":{
              "$ref":"#/definitions/ReservationReq"
            }
          }
        ],
        "responses":{
          "200":{
            "description":"Created",
            "schema":{
              "$ref":"#/definitions/ReservationRes"
            }
          },
          "400":{
            "description":"Bad Request"
          },
          "404":{
            "description":"Validation failed. Dates not available."
          },
          "406":{
            "description":"Validation failed."
          }
        }
      }
    },
    "/reservation/book-faulty":{
      "post":{
        "tags":[
          "reservation-controller"
        ],
        "summary":"Book Reservation",
        "description":"Books a reservation after all the validation criteria are met. Does not handle Race Condition.",
        "operationId":"bookFaultyUsingPOST",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"body",
            "name":"req",
            "description":"req",
            "required":true,
            "schema":{
              "$ref":"#/definitions/ReservationReq"
            }
          }
        ],
        "responses":{
          "200":{
            "description":"Created",
            "schema":{
              "$ref":"#/definitions/ReservationRes"
            }
          },
          "400":{
            "description":"Bad Request"
          },
          "404":{
            "description":"Validation failed. Dates not available."
          },
          "406":{
            "description":"Validation failed."
          }
        }
      }
    },
    "/reservation/cancel/{id}":{
      "delete":{
        "tags":[
          "reservation-controller"
        ],
        "summary":"Cancel Reservation",
        "description":"Cancels a reservation.",
        "operationId":"cancelUsingDELETE",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "name":"id",
            "in":"path",
            "description":"id",
            "required":true,
            "type":"integer",
            "format":"int64"
          }
        ],
        "responses":{
          "200":{
            "description":"Cancelled Reservation",
            "schema":{
              "$ref":"#/definitions/BasicRes"
            }
          },
          "404":{
            "description":"Not found"
          }
        }
      }
    },
    "/reservation/change/{id}":{
      "put":{
        "tags":[
          "reservation-controller"
        ],
        "summary":"Change Reservation",
        "description":"Allows you to update a reservation after all the validation criteria are met. Handle Race Condition.",
        "operationId":"changeUsingPUT",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "name":"id",
            "in":"path",
            "description":"id",
            "required":true,
            "type":"integer",
            "format":"int64"
          },
          {
            "in":"body",
            "name":"req",
            "description":"req",
            "required":true,
            "schema":{
              "$ref":"#/definitions/ReservationReq"
            }
          }
        ],
        "responses":{
          "200":{
            "description":"Created",
            "schema":{
              "$ref":"#/definitions/ReservationRes"
            }
          },
          "400":{
            "description":"Bad Request"
          },
          "404":{
            "description":"Validation failed. Reservation not found or Dates not available."
          },
          "406":{
            "description":"Validation failed. Or No change in reservation."
          }
        }
      }
    },
    "/reservation/list":{
      "get":{
        "tags":[
          "reservation-controller"
        ],
        "summary":"View Reservations",
        "description":"Displays all the reservations made.",
        "operationId":"listUsingGET",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "responses":{
          "200":{
            "description":"Success",
            "schema":{
              "type":"array",
              "items":{
                "$ref":"#/definitions/ReservationRes"
              }
            }
          }
        }
      }
    },
    "/reservation/lock":{
      "get":{
        "tags":[
          "reservation-controller"
        ],
        "summary":"Lock Reservation Booking or Changing",
        "description":"Locks the booking or changing of any reservation. Used to demonstrate that the Check Availability continues to work even though the Reservation Booking/Updating is locked.",
        "operationId":"lockUsingGET",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "responses":{
          "200":{
            "description":"OK",
            "schema":{
              "$ref":"#/definitions/BasicRes"
            }
          }
        }
      }
    },
    "/reservation/unlock":{
      "get":{
        "tags":[
          "reservation-controller"
        ],
        "summary":"Unlock Reservation Booking or Changing",
        "description":"Unlocks the booking or changing of any reservation.",
        "operationId":"unlockUsingGET",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "responses":{
          "200":{
            "description":"OK",
            "schema":{
              "$ref":"#/definitions/BasicRes"
            }
          }
        }
      }
    }
  },
  "definitions":{
    "ReservationRes":{
      "type":"object",
      "properties":{
        "arrivalDate":{
          "type":"string",
          "format":"date-time"
        },
        "departureDate":{
          "type":"string",
          "format":"date-time"
        },
        "email":{
          "type":"string"
        },
        "id":{
          "type":"integer",
          "format":"int64"
        },
        "name":{
          "type":"string"
        }
      }
    },
    "ReservationReq":{
      "type":"object",
      "properties":{
        "arrivalDate":{
          "type":"string",
          "format":"date-time"
        },
        "departureDate":{
          "type":"string",
          "format":"date-time"
        },
        "email":{
          "type":"string"
        },
        "id":{
          "type":"integer",
          "format":"int64"
        },
        "name":{
          "type":"string"
        }
      }
    },
    "AvailabilitySearchRes":{
      "type":"object",
      "properties":{
        "dates":{
          "type":"array",
          "items":{
            "type":"string",
            "format":"date-time"
          }
        },
        "end":{
          "type":"string",
          "format":"date-time"
        },
        "start":{
          "type":"string",
          "format":"date-time"
        }
      }
    },
    "BasicRes":{
      "type":"object",
      "properties":{
        "code":{
          "type":"string"
        },
        "message":{
          "type":"string"
        },
        "status":{
          "type":"boolean"
        }
      }
    }
  }
}